useCases/AssetMonitoringService/backend/requirements.txt
fastapi
uvicorn
pydantic-settings
requests
python-dotenv
